Se eliminan los archivos de seguimientos.php en managers de report_trackings  y groupal_trackings y se crean archivos de funciones comunes
